修正jdk1.8环境下出现的问题

+++ db/org.tinygroup.tinydb/src/test/java/org/tinygroup/tinydb/testcase/operator/TestSqlOperatorWithDialect.java
@@ -36,7 +36,7 @@ public class TestSqlOperatorWithDialect  extends BaseTest{
 		getOperator().batchInsert(insertBeans);
 		String sql = "select #{count(*)} as acount from ANIMAL";
 		Bean value =getOperator().getSingleValue(sql);
-		assertEquals(25, value.getProperty("acount"));
+		assertEquals(25, (int)value.getProperty("acount"));
 		
 		sql="select #{concat(name,'con')} as con from ANIMAL";
 		Bean[] beans= getOperator().getBeans(sql);
