+++ tinyflow/org.tinygroup.flow/src/test/java/org/tinygroup/flow/el/TestNextNodeByElInPageFlow.java
@@ -33,7 +33,7 @@ public class TestNextNodeByElInPageFlow extends AbstractFlowComponent{
 		c.put("a", 5);
 		c.put("b", 15);
 		pageFlowExecutor.execute("NextNodeByEl", c);
-		assertEquals(45, c.get("c"));
+		assertEquals(45, (int)c.get("c"));
 	}
 	
 	public void testNextNodeByE12(){
@@ -41,7 +41,7 @@ public class TestNextNodeByElInPageFlow extends AbstractFlowComponent{
 		c.put("a", 6);
 		c.put("b", 15);
 		pageFlowExecutor.execute("NextNodeByEl", c);
-		assertEquals(48, c.get("c"));
+		assertEquals(48, (int)c.get("c"));
 	}
 	
 	public void testFlowPropertyWithEl(){
@@ -52,7 +52,7 @@ public class TestNextNodeByElInPageFlow extends AbstractFlowComponent{
 		c.put("a", 5);
 		c.put("b", 15);
 		pageFlowExecutor.execute("NextNodeByEl", c);
-		assertEquals(45, c.get("c"));
+		assertEquals(45, (int)c.get("c"));
 	}
 	
 	public void testFlowPropertyWithEl2(){
@@ -63,7 +63,7 @@ public class TestNextNodeByElInPageFlow extends AbstractFlowComponent{
 		c.put("a", 6);
 		c.put("b", 15);
 		pageFlowExecutor.execute("NextNodeByEl", c);
-		assertEquals(48, c.get("c"));
+		assertEquals(48, (int)c.get("c"));
 	}
 	
 	//a+b = c 
@@ -73,7 +73,7 @@ public class TestNextNodeByElInPageFlow extends AbstractFlowComponent{
 		c.put("a", 5);
 		c.put("b", 15);
 		pageFlowExecutor.execute("NextNodeByNullEl", c);
-		assertEquals(35, c.get("c"));
+		assertEquals(35, (int)c.get("c"));
 	}
 	
 	public void testNextNodeByNullE12(){
@@ -81,6 +81,6 @@ public class TestNextNodeByElInPageFlow extends AbstractFlowComponent{
 		c.put("a", 6);
 		c.put("b", 15);
 		pageFlowExecutor.execute("NextNodeByNullEl", c);
-		assertEquals(36, c.get("c"));
+		assertEquals(36, (int)c.get("c"));
 	}
 }
