+++ tinyflow/org.tinygroup.flow/src/test/java/org/tinygroup/flow/test/testcase/TestSumFlow.java
@@ -32,10 +32,10 @@ public class TestSumFlow extends AbstractFlowComponent {
 		//sum = a + b
 		//sum = a + sum
 		flowExecutor.execute("testSumFlow", "begin", context);
-		assertEquals(4, context.get("sum"));
+		assertEquals(4, (int)context.get("sum"));
 		context.put("sum", 5);
 		flowExecutor.execute("testSumFlow", "sumComponent_1", context);
-		assertEquals(6, context.get("sum"));
+		assertEquals(6, (int)context.get("sum"));
 
 	}
 	
@@ -45,7 +45,7 @@ public class TestSumFlow extends AbstractFlowComponent {
 		context.put("b", 1);
 		context.put("sum", 5);
 		flowExecutor.execute("testSumFlow", "sumComponent_1", context);
-		assertEquals(6, context.get("sum"));
+		assertEquals(6, (int)context.get("sum"));
 
 	}
 	
@@ -55,7 +55,7 @@ public class TestSumFlow extends AbstractFlowComponent {
 		context.put("b", 2);
 		context.put("c", 5);
 		flowExecutor.execute("testSumFlow2",  context);
-		assertEquals(8, context.get("sum"));
+		assertEquals(8, (int)context.get("sum"));
 
 	}
 	
@@ -66,10 +66,10 @@ public class TestSumFlow extends AbstractFlowComponent {
 		//sum = a + b
 		//sum = a + sum
 		flowExecutor.execute("testSumFlowChild", "begin", context);
-		assertEquals(4, context.get("sum"));
+		assertEquals(4, (int)context.get("sum"));
 		context.put("sum", 5);
 		flowExecutor.execute("testSumFlowChild", "sumComponent_1", context);
-		assertEquals(6, context.get("sum"));
+		assertEquals(6, (int)context.get("sum"));
 
 	}
 	
@@ -80,9 +80,9 @@ public class TestSumFlow extends AbstractFlowComponent {
 		//sum = a + b
 		//sum = b + sum
 		flowExecutor.execute("testSumFlowChild2", "begin", context);
-		assertEquals(5, context.get("sum"));
+		assertEquals(5, (int)context.get("sum"));
 		flowExecutor.execute("testSumFlowChild2", "sumComponent_1", context);
-		assertEquals(7, context.get("sum"));
+		assertEquals(7, (int)context.get("sum"));
 
 	}
 
@@ -94,7 +94,7 @@ public class TestSumFlow extends AbstractFlowComponent {
 		//sum = b + sum
 		//sum = sum + sum
 		flowExecutor.execute("testSumFlowChild3", "begin", context);
-		assertEquals(10, context.get("sum"));
+		assertEquals(10, (int)context.get("sum"));
 
 	}
 	
@@ -103,7 +103,7 @@ public class TestSumFlow extends AbstractFlowComponent {
 		context.put("a", 1);
 		context.put("b", 2);
 		flowExecutor.execute("testSumFlowGrandson", "begin", context);
-		assertEquals(105, context.get("sum"));
+		assertEquals(105, (int)context.get("sum"));
 
 	}
 	
@@ -113,7 +113,7 @@ public class TestSumFlow extends AbstractFlowComponent {
 		context.put("a", 11);
 		context.put("b", 2);
 		flowExecutor.execute("testSumFlowEl", "begin", context);
-		assertEquals(24, context.get("sum"));
+		assertEquals(24, (int)context.get("sum"));
 
 	}
 	
@@ -122,7 +122,7 @@ public class TestSumFlow extends AbstractFlowComponent {
 		context.put("a", 2);
 		context.put("b", 12);
 		flowExecutor.execute("testSumFlowEl", "begin", context);
-		assertEquals(26, context.get("sum"));
+		assertEquals(26, (int)context.get("sum"));
 
 	}
